requirements.txt
streamlit
python-dotenv
fastapi
uvicorn
gunicorn
openai
firebase-admin